% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointPattern.R
\name{tool_exec}
\alias{tool_exec}
\title{Developing a script that enables a tool to run the point pattern analysis - estimating density and simulating point pattern.}
\usage{
tool_exec(in_params, out_params)
}
\arguments{
\item{in_params}{- input parameter containing the point feature.}

\item{out_params}{- output parameter containing the outputs for the estimated density and the simulated point pattern as the shapefiles.}
}
\value{
output_feature1 - \code{\link[spatstat]{density.ppp}} Output krige shapefile that contains the predictions of the values of unsampled locations from the Prediction_location dataset.

output_feature2 - \code{\link[spatstat]{simulate.ppm}} Output variance provides how far the values are deviated from the other and the mean and exports the output as a pdf file.
}
\description{
This script function converts the input point feature into a point pattern dataset and estimates the density of the converted point pattern
and simulates a point pattern based on the estimated density upon being run with the corresponding ArcGIS script tool.
}
\details{
This script function estimated the density of the point pattern using the kernel density estimation methodology using the kernel smoothed intensity function and simulates another point pattern by fitting a point pattern model.
}
\note{
The out_params parameter includes the return values of \emph{output_feature1}, which is the density and \emph{output_feature2}, which is the simulated point pattern.
}
\author{
Edzer Pebesma, Shankarlingam Sundaresan
}
\seealso{
\code{\link{spatstat}} , \code{\link{arcgisbinding}}
}

